+++ app/bundles/CoreBundle/Helper/BuilderTokenHelper.php
@@ -11,6 +11,7 @@
 
 namespace Mautic\CoreBundle\Helper;
 
+use Doctrine\DBAL\Query\Expression\CompositeExpression;
 use Doctrine\ORM\Query\Expr;
 use Mautic\CoreBundle\Factory\MauticFactory;
